requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
arize-phoenix>=0.0.6
openinference-semantic-conventions>=0.0.1
openinference-instrumentation-langchain>=0.0.1
openinference-instrumentation>=0.0.1
opentelemetry-sdk>=1.19.0
opentelemetry-api>=1.19.0
langchain>=0.3.1
langgraph>=0.3.1
langchain-openai>=0.3.1